🔨 🔥 Move start maya out of firm
- Fixed up lomax dependencies
- Remove input functions hack from avery
