+++ clients/lomax/lomax.nix
@@ -1,7 +1,5 @@
-{ pkgs, base, avery }:
+{ pkgs, base }:
 base.languages.rust.mkClient {
   name = "lomax";
   src = ./.;
-
-  manifestRsPath = "${avery.package.src}/src/manifest.rs";
 }
