--- functions/start-maya/.gitignore
@@ -1,4 +1,0 @@
-/target
-**/*.rs.bk
-nix-deps
-
