--- functions/start-maya/Cargo.toml
@@ -1,8 +1,0 @@
-[package]
-name = "start-maya"
-version = "0.1.0"
-authors = ["GBK pipeline  <pipeline@goodbyekansas.com>"]
-edition = "2018"
-
-[dependencies]
-gbk = { path = "nix-deps/gbk" }
