--- functions/start-maya/function.toml
@@ -1,14 +1,0 @@
-name = "start-maya"
-version = "0.1.0"
-
-[inputs]
-    [inputs.version]
-    type="string"
-    default="2019"
-
-[outputs]
-    [outputs.pid]
-    type="int"
-
-[execution-environment]
-type = "wasm"
