--- functions/start-maya/start-maya.nix
@@ -1,7 +1,0 @@
-{ pkgs, base, rustGbkUtils }:
-base.languages.rust.mkFunction {
-  manifest = ./function.toml;
-  name = "start-maya";
-  src = ./.;
-  rustDependencies = [ rustGbkUtils ];
-}
