+++ services/avery/avery.nix
@@ -1,9 +1,6 @@
-{ pkgs, base, inputFunctions }:
+{ pkgs, base }:
 with pkgs;
 base.languages.rust.mkService {
   name = "avery";
   src = ./.;
-
-  # TODO: This gives a nice workflow for now but should be removed later
-  inputFunctions = builtins.map (c: c.package) inputFunctions;
 }
