+++ services/avery/src/main.rs
@@ -1,20 +1,16 @@
 #![deny(warnings)]
 
-use std::{path::Path, sync::Arc};
+use std::sync::Arc;
 
-use futures;
-use slog::{error, info, o, Drain};
+use slog::{info, o, Drain};
 use slog_async;
 use slog_term;
 use structopt::StructOpt;
 use tonic::transport::Server;
 
 use avery::{
-    manifest::FunctionManifest,
     proto::{
-        functions_registry_server::{FunctionsRegistry, FunctionsRegistryServer},
-        functions_server::FunctionsServer,
-        RegisterRequest,
+        functions_registry_server::FunctionsRegistryServer, functions_server::FunctionsServer,
     },
     registry::FunctionsRegistryService,
     FunctionsService,
@@ -29,11 +25,7 @@ async fn ctrlc() {
 
 #[derive(StructOpt, Debug)]
 #[structopt(name = "avery")]
-struct AveryArgs {
-    /// function executor service address
-    #[structopt(short, long)]
-    skip_register_test_functions: bool,
-}
+struct AveryArgs {}
 
 // local server main loop
 #[tokio::main]
@@ -44,66 +36,11 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     let log = slog::Logger::root(drain, o!());
 
-    let args = AveryArgs::from_args();
+    let _args = AveryArgs::from_args();
 
     let port: u32 = 1939;
     let addr = format!("[::]:{}", port).parse().unwrap();
     let functions_registry_service = Arc::new(FunctionsRegistryService::new());
-    if !args.skip_register_test_functions {
-        // TODO: this is very temp but gives a nice workflow atm
-        info!(log, "registering functions from $inputFunctions...");
-        match std::env::var("inputFunctions") {
-            Ok(fnstring) => {
-                fnstring
-                    .split(';')
-                    .filter_map(|p| {
-                        let manifest_path = Path::new(p).join("manifest.toml");
-
-                        let manifest = FunctionManifest::parse(manifest_path)
-                            .map_err(|e| error!(log, "\"{}\". Skipping", e))
-                            .ok()?;
-
-                        let mut register_request = RegisterRequest::from(&manifest);
-                        let code_path = Path::new(p)
-                            .join("bin")
-                            .join(format!("{}.wasm", manifest.name()));
-                        info!(log, "reading code file from: {}", code_path.display());
-                        register_request.code = std::fs::read(code_path)
-                            .map_err(|e| {
-                                error!(
-                                    log,
-                                    "Failed to read code for function {}: {}. Skipping.",
-                                    manifest.name(),
-                                    e
-                                )
-                            })
-                            .ok()?
-                            .to_vec();
-                        Some(register_request)
-                    })
-                    .for_each(|f| {
-                        futures::executor::block_on(
-                            functions_registry_service.register(tonic::Request::new(f.clone())),
-                        )
-                        .map_or_else(
-                            |e| {
-                                error!(
-                                    log,
-                                    "Failed to register function \"{}\". Err: {}", f.name, e
-                                )
-                            },
-                            |_| (),
-                        );
-                    });
-            }
-            Err(e) => error!(
-                log,
-                "Tried to add functions from the env var but $inputFunctions was not set. {}", e
-            ),
-        };
-
-        info!(log, "done registering functions from $inputFunctions");
-    }
 
     let functions_service = FunctionsService::new(
         log.new(o!("service" => "functions")),
